+++ models/contacts.json
@@ -70,5 +70,17 @@
     "name": "Devon",
     "email": "devon@mail.com",
     "phone": "7043987993"
+  },
+  {
+    "id": "5721ab04-f9e1-4a8b-847c-163c1da85f8b",
+    "name": "Devon",
+    "email": "devon@mail.com",
+    "phone": "5555555555"
+  },
+  {
+    "id": "3c70cb29-64e5-413a-a75e-9fe88d1324cb",
+    "name": "Devon",
+    "email": "devon@mail.com",
+    "phone": "66666666666"
   }
 ]
