+++ package.json
@@ -12,6 +12,7 @@
     "cors": "2.8.5",
     "cross-env": "7.0.3",
     "express": "4.17.1",
+    "http-errors": "^2.0.0",
     "joi": "^17.7.0",
     "morgan": "1.10.0",
     "nanoid": "^4.0.0",
