+++ routes/api/contacts.js
@@ -1,5 +1,6 @@
 const express = require("express");
-const Joi = require("joi");
+const { contactSchema } = require("../../validation");
+const { BadRequest, NotFound } = require("http-errors");
 const {
   listContacts,
   getContactById,
@@ -10,23 +11,12 @@ const {
 
 const router = express.Router();
 
-const contactSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string()
-    .email({
-      minDomainSegments: 2,
-      tlds: { allow: ["com", "net"] },
-    })
-    .required(),
-  phone: Joi.string().required(),
-});
-
 router.get("/", async (req, res, next) => {
   try {
     const data = await listContacts();
     res.status(200).json({ data });
   } catch (err) {
-    res.status(404).json({ message: err.message });
+    throw new NotFound("Not found");
   }
 });
 
@@ -35,12 +25,11 @@ router.get("/:contactId", async (req, res, next) => {
     const contactId = req.params.contactId;
     const data = await getContactById(contactId);
     if (data === undefined) {
-      res.status(404).json({ message: "Not found" });
-      return;
+      throw new NotFound("Not found");
     }
     res.status(200).json({ data });
-  } catch (err) {
-    res.status(404).json({ message: "Not found" });
+  } catch (error) {
+    next(error);
   }
 });
 
@@ -48,13 +37,12 @@ router.post("/", async (req, res, next) => {
   try {
     const { error } = contactSchema.validate(req.body);
     if (error) {
-      res.status(400).json({ message: "missing required name field" });
-      return;
+      throw BadRequest("missing required name field");
     }
     const data = await addContact(req.body);
     res.status(201).json({ data });
-  } catch (err) {
-    res.status(500).json({ message: "Server error" });
+  } catch (error) {
+    next(error);
   }
 });
 
@@ -63,14 +51,13 @@ router.delete("/:contactId", async (req, res, next) => {
     const contactId = req.params.contactId;
     const data = await getContactById(contactId);
     if (!data) {
-      res.status(404).json({ message: "Not found" });
-      return;
+      throw new NotFound("Not found");
     }
     await removeContact(contactId);
 
     res.status(200).json({ message: "contact deleted" });
-  } catch (err) {
-    res.status(500).json({ message: "Server error" });
+  } catch (error) {
+    next(error);
   }
 });
 
@@ -78,20 +65,18 @@ router.put("/:contactId", async (req, res, next) => {
   try {
     const { error } = contactSchema.validate(req.body);
     if (error) {
-      res.status(400).json({ message: "missing required name field" });
-      return;
+      throw BadRequest("missing required name field");
     }
 
     const contactId = req.params.contactId;
 
     const updatedContact = await updateContact(contactId, req.body);
     if (!updatedContact) {
-      res.status(404).json({ message: "Not found" });
-      return;
+      throw new NotFound("Not found");
     }
     res.status(200).json({ updatedContact });
-  } catch (err) {
-    res.status(500).json({ message: "Server error" });
+  } catch (error) {
+    next(error);
   }
 });
